+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Renamed shard spaces to retention policies
 - Deprecated matching against regex in favor of explicit writing and querying on retention policies
+- Pure Go InfluxQL scanner
 
 ## v0.8.6 [2014-11-15]